+++ CHANGELOG.md
@@ -18,10 +18,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 - fixed SQL encoding whenever template is unresolved (https://github.com/ehrbase/ehrbase/pull/723)
+
+- modified handling of conflicting identified parties (https://github.com/ehrbase/ehrbase/issues/710)
+
 - Fixes wrong status code returned by EHRbase while creating FLAT composition ([#726](https://github.com/ehrbase/ehrbase/pull/726))
 
 - Fix NullPointerException while deleting unknown (or already deleted) composition
   parameter ([#722](https://github.com/ehrbase/ehrbase/pull/722)).
+
 
 ## [0.18.3]